Share sparse MLA fallback env handling
Move VLLM_TRITON_MLA_SPARSE and legacy SM120 alias parsing into a shared sparse_mla_env module. DeepSeek V4 attention and the SWA metadata builder now use the same helpers for reference fallback enablement, diagnostic dumps, and chunk-size knobs.

This preserves the current behavior where unset VLLM_TRITON_MLA_SPARSE auto-enables the fallback on SM12x, while an explicit new value overrides the legacy aliases.
